Add files via upload
The objective of developing an "English Dictionary App" is to collect generic book
information. Using JSON, create a text file to store, retrieve, and manipulate data.
Outcome: With the help of JSON, anyone can add and find entries from the dictionary,
along with their definitions. Removes the terms from the dictionary and closes the
window. The JSON functions open(), dump(), and load() are used to open, dump
(serialization), and load (Deserialization) the data.